move stl related codes to a subfolder
