+++ src/Microsoft.ML.TimeSeries/STL/BackendStructureBase.cs
