+++ src/Microsoft.ML.TimeSeries/STL/BasicParameters.cs
