+++ src/Microsoft.ML.TimeSeries/STL/LeastSquares.cs
