+++ src/Microsoft.ML.TimeSeries/STL/LocalRegression.cs
