+++ src/Microsoft.ML.TimeSeries/STL/MathsTools.cs
