+++ src/Microsoft.ML.TimeSeries/STL/SeasonalInfo.cs
